🧩 YD5J89 quality: split evaluator prepare coverage
