+++ .agentplane/tasks/202607221908-YD5J89/pr/meta.json
@@ -2,7 +2,7 @@
   "base": "main",
   "branch": "task/202607221908-YD5J89/migrate-context-and-evaluator-command-boundaries",
   "created_at": "2026-08-01T00:38:26.423Z",
-  "diffstat_sha256": "sha256:ac372608e3394d08ccd2d3b96191c3d33cb6fd027fe182336b5b93c311758e58",
+  "diffstat_sha256": "sha256:62f045b93eaae40605ec53b2dd5824f494c9864ee5f619bdd3eeb1f84e65cd1d",
   "last_verified_at": "2026-08-01T09:18:12.405Z",
   "last_verified_diffstat_sha256": "sha256:ac372608e3394d08ccd2d3b96191c3d33cb6fd027fe182336b5b93c311758e58",
   "pr_number": 4716,
